Rename device -> devices
