+++ pkg/devices/switches/switches_test.go
